src/main/python/model/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.3.1
ansiwrap==0.8.4
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backoff==1.10.0
backports.functools-lru-cache==1.6.4
beatrix-jupyterlab==2023.123.173907
beautifulsoup4==4.11.1
bleach==6.0.0
blessed==1.19.1
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-tpu-client==0.10
cloudpickle==2.2.1
cmdstanpy==0.9.5
colorama==0.4.6
colorful==0.5.5
conda==22.9.0
conda-content-trust==0+unknown
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
convertdate==2.4.0
cryptography==38.0.2
cycler==0.11.0
Cython==0.29.33
db-dtypes
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dm-tree==0.1.8
docker
entrypoints==0.4
ephem==4.1.4
fastapi==0.89.1
fastjsonschema==2.16.2
fbprophet==0.7.1
filelock==3.9.0
flit_core==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gcsfs==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
google-api-core==1.34.0
google-api-python-client==1.8.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-aiplatform==1.21.0
google-cloud-bigquery==3.4.2
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-monitoring==2.14.1
google-cloud-resource-manager==1.8.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpustat==1.0.0
greenlet==2.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.48.2
gym==0.23.1
gym-notices==0.0.8
h11==0.14.0
hijri-converter==2.2.4
holidays==0.19
htmlmin==0.1.12
httplib2==0.21.0
idna==3.4
ImageHash==4.3.1
imageio==2.25.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter-http-over-ws==0.0.8
jupyter-server==1.23.5
jupyter-server-mathjax==0.2.6
jupyter-server-proxy==3.2.2
jupyterlab==3.4.8
jupyterlab-git==0.41.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
jupyterlab-widgets==3.0.5
jupytext==1.14.4
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
kubernetes==25.3.0
llvmlite==0.39.1
LunarCalendar==0.0.9
lz4==4.3.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
msgpack==1.0.4
multidict==6.0.4
multimethod==1.9.1
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbdime==3.1.1
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook-executor==0.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.21.6
nvidia-ml-py==11.495.46
oauth2client==4.1.3
oauthlib==3.2.2
opencensus==0.11.1
opencensus-context==0.1.3
opentelemetry-api==1.1.0
opentelemetry-exporter-otlp==1.1.0
opentelemetry-exporter-otlp-proto-grpc==1.1.0
opentelemetry-proto==1.1.0
opentelemetry-sdk==1.1.0
opentelemetry-semantic-conventions==0.20b0
packaging
pandas==1.3.5
pandas-profiling==3.6.2
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.13.0
pmdarima==2.0.2
prettytable==0.7.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
py-spy==0.3.14
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.4
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
PyMeeus==0.5.12
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pystan==2.19.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
ray==2.2.0
ray-cpp==2.2.0
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.2.0
rsa==4.9
ruamel-yaml-conda==0.15.100
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==66.1.1
setuptools-git==1.2
Shapely==1.8.5.post1
simpervisor==0.4
six==1.16.0
sktime==0.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.22.0
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
tensorboardX==2.5.1
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typeguard==2.13.3
typer==0.7.0
typing_extensions==4.4.0
uritemplate==3.0.1
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
visions==0.7.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.14.1
yarl==1.8.2
zipp==3.11.0
zstandard==0.18.0
